NotionLinker/requirements.txt

notion-client>=2.3.0
python-dateutil>=2.9.0
python-dotenv>=1.0.1
pathlib2>=2.3.0; python_version < "3.4"
